sync RMF to 0fdafa08866d0f5f1378d33354546deaae892a03
git-svn-id: https://svn.salilab.org/imp/trunk@16766 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ modules/rmf/dependency/RMF/src/backend/avro/AvroSharedData.impl.h
@@ -64,7 +64,7 @@ int AvroSharedData<Base>::add_child(int node, std::string name, int t) {
   P::access_node(index).type = boost::lexical_cast<std::string>(NodeType(t));
   add_child(node, index);
   P::add_node_key();
-  RMF_INTERNAL_CHECK(get_type(index) == t,
+  RMF_INTERNAL_CHECK(get_type(index) == static_cast<unsigned int>(t),
                      internal::get_error_message("Types don't match for node ",
                                        name, ": ", NodeType(t), " (", t,
                                        ") vs ",
